# Implement IP.String without relying on the stdlib.