@danderson 2020-12-30
For comparison, net.IP.String on the same inputs and machine:

```
name                       time/op
IPStringStd/v4-8           65.8ns
IPStringStd/v6-8            251ns
IPStringStd/v6_ellipsis-8   184ns
IPStringStd/v6_v4-8        71.1ns
```

Number of allocs and alloc size is the same as netaddr.IP.String. v6_zone doesn't parse in stdlib, hence why it's missing from the list.